Merchant_revenue_by_date
"merchant_revenue_by_date" according to the spec should return a dollar amount.
The spec harness currently has this method returning an array of merchants

Also running the date at "2011-02-27" seems to return 0 because there were no invoices "created_at" on this date. Although there are several invoices "updated_at" this date. We figured that revenue would be booked on the day it was created. Let us know if we are off
Thanks - Allan / Brennan